+++ app/Http/Controllers/ServiceCategoriesController.php
@@ -15,6 +15,7 @@ class ServiceCategoriesController extends Controller
     public function index()
     {
         //
+        $this->authorize('viewAny', Service_categories::class);
         $service_categories = Service_categories::all();
         return view('service_categories.index')->with('service_categories', $service_categories)->with('scripts',true);
     }
@@ -39,6 +40,7 @@ class ServiceCategoriesController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create', Service_categories::class);
         $validateData =$request->validate(
             ["nama_layanan"=>"required"]
         );
@@ -84,6 +86,7 @@ class ServiceCategoriesController extends Controller
     public function update(Request $request, Service_categories $service_category)
     {
         //
+        $this->authorize('update', Service_categories::class);
         $validateData = $request->validate(
             ['nama_layanan'=>'required']
         );
@@ -102,6 +105,7 @@ class ServiceCategoriesController extends Controller
     public function destroy(Service_categories $service_category)
     {
         //
+        $this->authorize('delete', Service_categories::class);
         $service_category->delete();
         return redirect()->route('service_categories.index')->with('success', "Data kategori layanan $service_category->nama_layanan berhasil dihapus");
     }
