+++ app/Http/Controllers/StudyProgramsController.php
@@ -15,6 +15,7 @@ class StudyProgramsController extends Controller
     public function index()
     {
         //
+        $this->authorize('viewAny', Study_programs::class);
         $study_programs = study_programs::all();
         return view('study_programs.index')->with('study_programs', $study_programs)->with('scripts',true);
     }
@@ -39,6 +40,7 @@ class StudyProgramsController extends Controller
     public function store(Request $request)
     {
         //
+        $this->authorize('create', Study_programs::class);
         $validateData = $request->validate([
             'nama_prodi' => 'required'
         ]);
@@ -84,6 +86,7 @@ class StudyProgramsController extends Controller
     public function update(Request $request, Study_programs $study_program)
     {
         //
+        $this->authorize('update', Study_programs::class);
         $validateData =$request->validate([
             'nama_prodi' => 'required'
         ]);
@@ -102,6 +105,7 @@ class StudyProgramsController extends Controller
     public function destroy(Study_programs $study_program)
     {
         //
+        $this->authorize('delete', Study_programs::class);
         $study_program->delete();
         return redirect()->route('study_programs.index')->with('success', "Data Program Studi $study_program->nama_prodi berhasil dihapus");
     }
